Use u32 as the compressed state key
This gives further speed up at the cost of a reasonable assumption that,
in a single block, less than 4 billion unique keys are accessed.

+++ execution/transaction-orderer/src/transaction_compressor.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 use std::hash::Hash;
 use std::rc::Rc;
 
-pub type CompressedKey = u64;
+pub type CompressedKey = u32;
 
 #[derive(Default)]
 pub struct CompressedPTransactionInner<T> {
